Inline AsAcceptOnlyInterface test data into tests
