--- Source/Moq.Analyzers.Test/Data/AsAcceptOnlyInterface.TestBadAsForAbstractClass.cs
@@ -1,17 +1,0 @@
-﻿using Moq;
-
-namespace AsAcceptOnlyInterface.TestBadAsForAbstractClass;
-
-public abstract class BaseSampleClass
-{
-    public int Calculate() => 0;
-}
-
-internal class MyUnitTests
-{
-    private void TestBadAsForAbstractClass()
-    {
-        var mock = new Mock<BaseSampleClass>();
-        mock.As<BaseSampleClass>();
-    }
-}
